+++ src/Service/Indexing/IndexingLocations.php
@@ -18,6 +18,7 @@ final class IndexingLocations extends AbstractIndexingElastic
     public function __construct(
         private readonly SerializerInterface $serializer,
         private readonly Client $client,
+        private readonly string $viewTimezone,
     ) {
         parent::__construct($this->client);
     }
@@ -28,7 +29,7 @@ final class IndexingLocations extends AbstractIndexingElastic
             ->withGroups([IndexNames::Locations->value]);
         $contextBuilder = (new DateTimeNormalizerContextBuilder())
             ->withContext($contextBuilder)
-            ->withTimezone('Europe/Copenhagen')
+            ->withTimezone($this->viewTimezone)
             ->withFormat(IndexFieldTypes::DATEFORMAT);
         $data = $this->serializer->normalize($item, null, $contextBuilder->toArray());